+++ CHANGELOG.md
@@ -47,6 +47,8 @@ changes.
 
 - Replaced existing websocket server with production-grade one
 
+- Removed `Greetings` messages from hydra-node history
+
 ## [0.9.0] - 2023-03-02
 
 :dragon_face: Renamed the repository from `hydra-poc` to [`hydra`](https://github.com/input-output-hk/hydra)!